+++ backend/src/data/data.sql
@@ -2,7 +2,7 @@
 CREATE TABLE IF NOT EXISTS books (
     id SERIAL PRIMARY KEY,
     title VARCHAR(255) NOT NULL,
-    price INT NOT NULL,
+    price DECIMAL NOT NULL,
     quantity INT NOT NULL,
     author VARCHAR(255) NOT NULL,
     genre VARCHAR(100),
